Refactor: more readable naming Fix: navigation in main_nav_graph.xml
