+++ app/src/main/java/com/velord/composemultiplebackstackdemo/ui/main/bottomNavigation/left/LeftGraphFragment.kt
@@ -18,7 +18,7 @@ class LeftGraphFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View = setContentWithTheme {
         AddNewScreen(R.string.add_new_screen_left) {
-            findNavController().navigate(R.id.toInDevelopmentFragment)
+            findNavController().navigate(R.id.toInDevelopmentFragmentFromLeftGraphFragment)
         }
     }
 }
