--- app/src/main/java/com/velord/composemultiplebackstackdemo/ui/main/inDevelopment/TestFragment.kt
@@ -1,21 +1,0 @@
-package com.velord.composemultiplebackstackdemo.ui.main.inDevelopment
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-import com.velord.composemultiplebackstackdemo.R
-import com.velord.composemultiplebackstackdemo.ui.compose.screen.AddNewScreen
-import com.velord.composemultiplebackstackdemo.ui.compose.theme.setContentWithTheme
-
-class TestFragment : Fragment() {
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View = setContentWithTheme {
-        AddNewScreen(R.string.app_name)
-    }
-}
