+++ app/src/main/java/com/velord/composemultiplebackstackdemo/ui/main/overlay/OverlayViewModel.kt
@@ -1,4 +1,4 @@
-package com.velord.composemultiplebackstackdemo.ui.main.inDevelopment
+package com.velord.composemultiplebackstackdemo.ui.main.overlay
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
@@ -7,7 +7,7 @@ import com.velord.composemultiplebackstackdemo.ui.navigation.NavigationData
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.launch
 
-class InDevelopmentViewModel : ViewModel() {
+class OverlayViewModel : ViewModel() {
 
     val navigationEvent = MutableSharedFlow<NavigationData>()
