+++ app/src/main/res/navigation/left_nav_graph.xml
@@ -10,7 +10,7 @@
         android:label="LeftGraphFragment" >
 
         <action
-            android:id="@+id/toInDevelopmentFragment"
+            android:id="@+id/toInDevelopmentFragmentFromLeftGraphFragment"
             app:destination="@id/inDevelopmentFragment"
             app:enterAnim="@anim/slide_in_up"
             app:popExitAnim="@anim/slide_in_down"
@@ -20,7 +20,7 @@
 
     <fragment
         android:id="@+id/inDevelopmentFragment"
-        android:name="com.velord.composemultiplebackstackdemo.ui.main.inDevelopment.InDevelopmentFragment"
+        android:name="com.velord.composemultiplebackstackdemo.ui.main.overlay.OverlayFragment"
         android:label="InDevelopmentFragment" />
 
 </navigation>
