+++ app/src/main/res/navigation/right_nav_graph.xml
@@ -10,13 +10,13 @@
         android:label="RightGraphFragment" >
 
         <action
-            android:id="@+id/toInDevelopmentFragment"
+            android:id="@+id/toInDevelopmentFragmentFromRightGraphFragment"
             app:destination="@id/inDevelopmentFragment" />
     </fragment>
 
     <fragment
         android:id="@+id/inDevelopmentFragment"
-        android:name="com.velord.composemultiplebackstackdemo.ui.main.inDevelopment.InDevelopmentFragment"
+        android:name="com.velord.composemultiplebackstackdemo.ui.main.overlay.OverlayFragment"
         android:label="InDevelopmentFragment" />
 
 </navigation>
